refactor: move non-route export 2
