+++ app/routes/decks/$id/edit.tsx
@@ -1,7 +1,6 @@
 import { useTinyForm } from "@hiogawa/tiny-form/dist/react";
 import { useNavigate } from "@remix-run/react";
 import { useMutation } from "@tanstack/react-query";
-import { DeckNavBarMenuComponent } from ".";
 import { $R, R } from "../../../misc/routes";
 import { rpcClientQuery } from "../../../trpc/client";
 import { asNumberInput } from "../../../utils/form-utils";
@@ -11,6 +10,7 @@ import { cls } from "../../../utils/misc";
 import type { PageHandle } from "../../../utils/page-handle";
 import { toast } from "../../../utils/toast-utils";
 import type { LoaderData } from "./_utils.server";
+import { DeckNavBarMenuComponent } from "./index.utils";
 export { loader } from "./_utils.server";
 
 export const handle: PageHandle = {
