+++ app/routes/decks/$id/history-graph.tsx
@@ -2,7 +2,6 @@ import { Transition } from "@hiogawa/tiny-transition/dist/react";
 import { useQuery } from "@tanstack/react-query";
 import type { ECharts } from "echarts";
 import React from "react";
-import { DeckNavBarMenuComponent } from ".";
 import { SelectWrapper, transitionProps } from "../../../components/misc";
 import {
   EchartsComponent,
@@ -20,8 +19,9 @@ import {
 import { cls } from "../../../utils/misc";
 import type { PageHandle } from "../../../utils/page-handle";
 import { formatDateRange } from "../../../utils/temporal-utils";
-
 import type { LoaderData } from "./_utils.server";
+
+import { DeckNavBarMenuComponent } from "./index.utils";
 export { loader } from "./_utils.server";
 
 export const shouldRevalidate = disableUrlQueryRevalidation;
