+++ app/routes/decks/$id/practice.tsx
@@ -20,7 +20,8 @@ import type { PageHandle } from "../../../utils/page-handle";
 import { BookmarkEntryComponent } from "../../bookmarks";
 
 import type { LoaderData } from "./_utils.server";
-import { DeckNavBarMenuComponent, QueueStatisticsComponent } from "./index";
+import { QueueStatisticsComponent } from "./index";
+import { DeckNavBarMenuComponent } from "./index.utils";
 export { loader } from "./_utils.server";
 
 export const handle: PageHandle = {
